[MOV] hr_fleet: Split model files according to coding guidelines
Applies the same changes as 6747402a96d3bd6013aa9e0ec29bca17c8ffc4a9 to
hr_fleet

Task ID: 2415309
