+++ addons/hr_fleet/models/__init__.py
@@ -6,4 +6,6 @@ from . import res_users
 from . import fleet_vehicle_assignation_log
 from . import fleet_category
 from . import fleet_vehicle
-from . import fleet_vehicle_cost
+from . import fleet_vehicle_log_contract
+from . import fleet_vehicle_log_services
+from . import fleet_vehicle_odometer
