+++ addons/hr_fleet/models/fleet_vehicle_log_services.py
@@ -3,15 +3,6 @@
 
 from odoo import api, fields, models
 
-class FleetVehicleLogContract(models.Model):
-    _inherit = 'fleet.vehicle.log.contract'
-
-    purchaser_employee_id = fields.Many2one(
-        related='vehicle_id.driver_employee_id',
-        string='Current Driver (Employee)',
-    )
-    fleet_is_internal = fields.Boolean(related='vehicle_id.fleet_is_internal')
-
 class FleetVehicleLogServices(models.Model):
     _inherit = 'fleet.vehicle.log.services'
