+++ addons/hr_fleet/models/fleet_vehicle_odometer.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import fields, models
+
+
+class FleetVehicleOdometer(models.Model):
+    _inherit = 'fleet.vehicle.odometer'
+
+    driver_employee_id = fields.Many2one(
+        related='vehicle_id.driver_employee_id', string='Current Driver (Employee)',
+        readonly=True,
+    )
